pwm: Convert period and duty to unsigned int
We don't need a 64 bits value to store nanoseconds

Discused with:	ian, jhibbits

+++ usr.sbin/pwm/pwm.c
@@ -190,9 +190,9 @@ main(int argc, char *argv[])
 		}
 		break;
 	case PWM_SHOW_CONFIG:
-		printf("period: %ju\nduty: %ju\nenabled:%d\n",
-		    (uintmax_t) state.period,
-		    (uintmax_t) state.duty,
+		printf("period: %u\nduty: %u\nenabled:%d\n",
+		    state.period,
+		    state.duty,
 		    state.enable);
 		break;
 	case PWM_PERIOD:
